delete check chunk
This chunk generates an error with "Submit Answer". You could use a check chunk perhaps with the following code snippet:

	grade_this({
		if (identical(.user_code, .solution_code)) {
			pass("Success!")
		}
		fail()
	})

But in that case the exercise and solution have to be identical, meaning that you would repeat in the solution (or delete in the exercise chunk) the commented line "# update the role for animal_name and animal_type".

As you never used check chunks I suggest to delete this check chunk here.